refactor: useModal 커스텀훅으로 모달 상태 관리
