+++ src/hooks/useModal.ts
@@ -1,0 +1,17 @@
+import { useState } from 'react';
+
+const useModal = () => {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleOpen = () => {
+    setIsOpen(true);
+  };
+
+  const handleClose = () => {
+    setIsOpen(false);
+  };
+
+  return { isOpen, handleOpen, handleClose };
+};
+
+export default useModal;
